Switch enlargen/close all windows
I realized that o is better for fast toggling which could be a  nice
feature to add to my workflow in order to fastly focus on one window
and then return to the previous layout. Capital O makes more sense
since it has more impact if  accidently executed as well.