Major refactor of configuration
Made the configuration more structured by sectioning the different
settings depending on the package that they relate to.

Removed the code related to automatically switching language when using
Swedish/English. It's been removed for a while now and I haven't missed
it so it's better to add it when actually needed.