Switch from mu4e to custom notmuch setup
Changing from the built in mu4e module to a custom notmuch based email
module. This module is based on the work of xfu and is hence very
customized. I will take time to configure it more myself but this is a
first attempt.

Due to all the customization's being so large I thought it would be
better to put it in it's own module.