Add the org-brain package
Playing around with org-brain to figure out how it works and if I like
it or not :). I am a bit split because I have grown so accustomed to be
able to easily search with Swiper in my notebook. However I like the
concept so let's give it a try :)