Add section for disabling Flycheck in modes
Some modes I don't have any flycheck available for. At the moment
Flycheck is activated globally and in order to turn it off for some
modes I added this section.

The pro of doing it this way rather than the reverse is that I can get a
good overview of which modes I don't have it in rather than only
activating it for the ones I know since I might accidentally miss ones
that I have for then.