Remove password-store as auth-source
I don't seem to get the password-store to work as an auth-source. In the
meantime I can just get by with a .gpg encrypted file with default
location inside the doom directory.

I only use it for Magithub and since that needs a computer specific
token anyhow that is not really a big deal.

I should add that as something to investigate in the future.