Change d in eshell to a function instead of alias
I started using the dired-quick-sort package and suddenly the alias in
eshell for dired behaved differently. I haven't checked the details but
I suspect that it might be related to that dired-quick-sort needs to
pass filter flags to ls. This input is optional so I can't just press d
without specifiying a directory, which is a bit redundant when I want
the current dir. Implementing this function instead seems to solve the
issue.