Override lazy-loading because of daemon setup
Lazy loading makes a lot of sense when running single Emacs instances.
When experimenting with running Emacs as a daemon and starting frames
that attaches to the server the benefits of lazy loading is not that
obvious.

It make more sense in daemon mode to load most of what I know I want on
startup. That way when I open a specific Emacs frame it will open
directly. Not hitting the lazy-loading and slowing down say the first
eshell frame, whilst the next one will be quick.